Added functionality that allows massive import of subdomains
